refactor: organize components into directories
