+++ .vscode/settings.json
@@ -7,5 +7,5 @@
   ],
   // https://stackoverflow.com/a/64573900/8238129
   "vetur.validation.template": false,
-  "typescript.tsserver.useSyntaxServer": "never"
+  "javascript.validate.enable": false
 }
