+++ components/chart/BarChart.vue
