+++ components/chart/TrackerGraph.vue
