+++ components/share-button/ShareButtonTwitter.vue
