+++ components/the-selector/TheSelectorExample.vue
