+++ components/the-selector/TheSelectorSparql.vue
