+++ components/unit/file-explorer/viewer/UnitFileExplorerViewerCsv.vue
@@ -8,20 +8,11 @@
 </template>
 
 <script>
-import FileManager from '~/utils/file-manager'
+import mixin from './mixin'
 
 export default {
   name: 'UnitFileExplorerViewerCsv',
-  props: {
-    filename: {
-      type: String,
-      required: true
-    },
-    fileManager: {
-      type: FileManager,
-      required: true
-    }
-  },
+  mixins: [mixin],
   data() {
     return {
       csvText: '',
