+++ components/unit/file-explorer/viewer/UnitFileExplorerViewerHtml.vue
@@ -1,0 +1,13 @@
+<template>
+  <iframe :src="path" width="100%" height="500px" />
+</template>
+
+<script>
+import mixin from './mixin'
+import mixinPath from './mixin-path'
+
+export default {
+  name: 'UnitFileExplorerViewerHtml',
+  mixins: [mixin, mixinPath]
+}
+</script>
