+++ components/unit/file-explorer/viewer/UnitFileExplorerViewerImage.vue
@@ -1,0 +1,13 @@
+<template>
+  <img width="100%" :src="path" alt="problem" />
+</template>
+
+<script>
+import mixin from './mixin'
+import mixinPath from './mixin-path'
+
+export default {
+  name: 'UnitFileExplorerViewerImage',
+  mixins: [mixin, mixinPath]
+}
+</script>
