+++ components/unit/file-explorer/viewer/mixin-path.js
@@ -1,0 +1,7 @@
+export default {
+  computed: {
+    path() {
+      return URL.createObjectURL(this.fileManager.fileDict[this.filename])
+    }
+  }
+}
