+++ components/unit/file-explorer/viewer/mixin.js
@@ -1,0 +1,14 @@
+import FileManager from '@/utils/file-manager'
+
+export default {
+  props: {
+    fileManager: {
+      type: FileManager,
+      required: true
+    },
+    filename: {
+      type: String,
+      required: true
+    }
+  }
+}
