+++ jsconfig.json
@@ -3,8 +3,22 @@
     "baseUrl": ".",
     "paths": {
       "~/*": ["./*"],
-      "@/*": ["./*"]
+      "@/*": ["./*"],
+      "~~/*": ["./*"],
+      "@@/*": ["./*"]
     }
   },
-  "exclude": ["node_modules", ".nuxt", "dist"]
+  "exclude": [
+    ".github",
+    ".husky",
+    ".nuxt",
+    ".vscode",
+    "assets",
+    "config",
+    "dist",
+    "manifests/icons",
+    "node_modules",
+    "static",
+    "vega-examples",
+  ]
 }
